whatsapp_qrljacker/requirements.txt
streamlit==1.38.0 
flask==3.0.3 
pyzbar==0.1.9 
pillow==10.4.0 s
elenium==4.25.0 
pyyaml==6.0.2